# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This file for generate swagger",
        "title": "Registration API",
        "contact": {}
    },
    "paths": {
        "/confirm": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Confirm registration student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Confirm registration student",
                "operationId": "Confirm registration student",
                "parameters": [
                    {
                        "description": "Confirm registration student",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registration.ConfirmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration confirmed successfully."
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPBadRequestErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/submit": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Submit registration student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Submit registration student",
                "operationId": "Submit registration student",
                "parameters": [
                    {
                        "description": "Submit registration student",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registration.ConfirmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration submitted successfully."
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPBadRequestErrors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httputil.ErrorCode": {
            "type": "string",
            "enum": [
                "0001",
                "0002",
                "0003",
                "0004",
                "0005",
                "0006",
                "9999"
            ],
            "x-enum-varnames": [
                "BadRequest",
                "NotFound",
                "CreateUserError",
                "StatusForbidden",
                "StatusUnauthorized",
                "InternalServerError",
                "InvalidStatus"
            ]
        },
        "httputil.HTTPBadRequestError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httputil.HTTPBadRequestErrors": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/httputil.ErrorCode"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httputil.HTTPBadRequestError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/httputil.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "messageEn": {
                    "type": "string"
                }
            }
        },
        "registration.ConfirmCourseRequest": {
            "type": "object",
            "required": [
                "action",
                "classid",
                "coursecode",
                "courseid",
                "creditattempt",
                "grademode",
                "section"
            ],
            "properties": {
                "action": {
                    "type": "integer"
                },
                "classid": {
                    "type": "integer"
                },
                "coursecode": {
                    "type": "string"
                },
                "courseid": {
                    "type": "integer"
                },
                "creditattempt": {
                    "type": "number"
                },
                "grademode": {
                    "type": "string",
                    "enum": [
                        "GD",
                        "SW"
                    ]
                },
                "section": {
                    "type": "string"
                }
            }
        },
...
        "registration.ConfirmRequest": {
            "type": "object",
            "required": [
                "course",
                "ip"
            ],
            "properties": {
                "course": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/registration.ConfirmCourseRequest"
                    }
                },
                "ip": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}